# Remove duplicate work