--- a/bignum.c
+++ b/bignum.c
@@ -5,27 +5,40 @@ void solve_bignum(int sides, mpz_t result) {
     mpz_set_ui(result, 0);
     mpz_t adder;
     mpz_init(adder);
+    int use_rows = sides / 2 > 1 ? sides / 2 : 1;
+    int use_cols = sides / 2 > 1 ? sides / 2 : 1;
     bool sides_even = (sides % 2 == 0);
-    int use_sides = sides;
+    bool sides_large_odd = (!sides_even && sides > 1);
+    if(sides_large_odd){
+      use_rows += 1;
+      use_cols += 1;
+    }
 
     printf("Solving with bignums for %dx%d grid...\n", sides, sides);
 
-    if(sides_even) {
-        use_sides = sides / 2;
-    }
-    for(int row = 0; row < use_sides; row++) {
-        for(int col = 0; col < use_sides; col++) {
+    for(int row = 0; row < use_rows; row++) {
+        for(int col = 0; col < use_cols; col++) {
+            if(sides_large_odd && (row == (use_rows-1) && col != (use_cols-1))) {
+                continue;
+            }
             bool* visited = calloc(sides*sides, sizeof(bool));
             mpz_set_ui(adder, 1);
             solve_bignum_recursive(row, col, sides, visited, adder);
-            mpz_add(result, result, adder);
             free(visited);
             printf("\tSolved square at row %d, column %d: %s\n", row+1, col+1, mpz_get_str(NULL, 10, adder));
+            if(sides_even) {
+                printf("\t\tEven sides; square will be used four times\n");
+                mpz_mul_ui(adder, adder, 4);
+            } else if(sides_large_odd) {
+                if(row != use_rows-1 || col != use_cols-1) {
+                    printf("\t\tOdd sides; square will be used four times\n");
+                    mpz_mul_ui(adder, adder, 4);
+                } else {
+                    printf("\t\tCenter square is only used once\n");
+                }
+            }
+            mpz_add(result, result, adder);
         }
-    }
-    if(sides_even) {
-	printf("\tEven sides detected: multiplying by four\n");
-	mpz_mul_ui(result, result, 4);
     }
     mpz_clear(adder);
     return;