--- a/solver.c
+++ b/solver.c
@@ -2,13 +2,20 @@
 
 unsigned long long solve(int sides) {
     unsigned long long sum = 0;
+    int use_sides = sides;
+    if(sides % 2 == 0) {
+        use_sides = sides / 2;
+    }
     printf("Solving for %dx%d grid...\n", sides, sides);
-    for(int row = 0; row < sides; row++) {
-        for(int col = 0; col < sides; col++) {
+    for(int row = 0; row < use_sides; row++) {
+        for(int col = 0; col < use_sides; col++) {
             bool* visited = calloc(sides*sides, sizeof(bool));
             sum += solve_recursive(row, col, sides, visited, 1);
             free(visited);
         }
+    }
+    if(sides % 2 == 0) {
+        sum *= 4;
     }
     return(sum);
 }