```
fsf-home boggle_paths remove_duplicate_work % echo "Master:\n"; git checkout --quiet master; make > /dev/null; time ./boggle_paths 4; echo "\nOptimiz
Master:

Solving for 4x4 grid...
A board with 4 sides has 12029640 possible solutions
./boggle_paths 4  1.96s user 0.01s system 99% cpu 1.974 total

Optimized:

Solving for 4x4 grid...
	Solved square at row 1, column 1: 1012520
		Even sides; square will be used four times
	Solved square at row 1, column 2: 740111
		Even sides; square will be used four times
	Solved square at row 2, column 1: 740111
		Even sides; square will be used four times
	Solved square at row 2, column 2: 514668
		Even sides; square will be used four times
A board with 4 sides has 12029640 possible solutions
./boggle_paths 4  0.51s user 0.00s system 99% cpu 0.521 total
```